+++ src/main/java/org/yez/petbackend/service/transaction/CreateTransactionDto.java
@@ -1,6 +1,5 @@
 package org.yez.petbackend.service.transaction;
 
-import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 import org.yez.petbackend.domain.transaction.TransactionType;
 
@@ -11,7 +10,7 @@ import java.util.UUID;
 public record CreateTransactionDto(
         UUID creatorId,
         UUID groupId,
-        float amount,
+        double amount,
         String description,
         @Nullable TransactionType type,
         @Nullable Instant transactionTime,
